refactor: split motion, projectiles, ship control systems
