+++ src/game/systems/projectiles/index.ts
@@ -1,0 +1,4 @@
+export { advanceProjectiles } from './advance.js';
+export { fireProjectile } from './spawn.js';
+export { TEMP_POS } from './sharedTemps.js';
+export type { FireProjectileOptions, FireProjectileOverride } from './spawn.js';
