+++ src/game/systems/projectiles/sharedTemps.ts
@@ -1,0 +1,4 @@
+import { Vector3 } from 'three';
+
+export const TEMP_POS = new Vector3();
+export const TEMP_TARGET = new Vector3();
