+++ src/game/systems/shipControl/sharedTemps.ts
@@ -1,0 +1,6 @@
+import { Quaternion, Vector3 } from 'three';
+
+export const TEMP_DIR = new Vector3();
+export const TEMP_POS = new Vector3();
+export const TEMP_REL_POS = new Vector3();
+export const TEMP_QUAT = new Quaternion();
