+++ src/game/systems/shipControl/weapons.ts
@@ -1,0 +1,46 @@
+import { Vector3 } from 'three';
+import type { GameState, ShipEntity } from '../../../types/index.js';
+import { recordShotMetrics } from '../../metrics.js';
+import { fireProjectile } from '../projectiles.js';
+import type { ShipDecision } from './aiExecutor.js';
+import { TEMP_DIR } from './sharedTemps.js';
+
+export function handleShipWeapons(
+  state: GameState,
+  ship: ShipEntity,
+  decision: ShipDecision,
+  preferredTarget: ShipEntity | null,
+): void {
+  if (!decision.firePrimary || ship.ship.cooldown > 0) {
+    return;
+  }
+
+  (ship.muzzleFlashes ??= []).push({
+    local: new Vector3(0, 0, ship.transform.scale * 1.6),
+    t0: state.time,
+    amp: 1,
+    bulletType: ship.ship.bulletType,
+  });
+
+  const fireDir = TEMP_DIR.copy(decision.heading);
+  if (fireDir.lengthSq() < 1e-5) fireDir.set(0, 0, 1);
+  else fireDir.normalize();
+
+  const distanceToTarget = preferredTarget
+    ? ship.transform.position.distanceTo(preferredTarget.transform.position)
+    : undefined;
+  const deltaY = preferredTarget
+    ? preferredTarget.transform.position.y - ship.transform.position.y
+    : undefined;
+
+  recordShotMetrics(state.ai.metrics, {
+    shipId: ship.id,
+    hull: ship.ship.hull,
+    time: state.time,
+    distance: distanceToTarget,
+    deltaY,
+  });
+
+  fireProjectile(state, ship, fireDir);
+  ship.ship.cooldown = ship.ship.fireRate;
+}
